Extend API of jpeg_encoder with encode_components methods (#332)
Add an additional method for advanced encodings scenarios. This methods allow to encode components with different
coding parameters like: near lossless, jpeg-ls preset coding parameters or scans with different interleave modes.
Extend the decoder to also be able to decode scans with different interleave modes.

+++ src/constants.hpp
@@ -13,7 +13,7 @@ constexpr int32_t default_reset_threshold{64}; // Default RESET value as defined
 
 constexpr int32_t minimum_component_count{1};
 constexpr int32_t maximum_component_count{255};
-constexpr size_t maximum_component_count_in_scan{4};
+constexpr int32_t maximum_component_count_in_scan{4};
 constexpr int32_t minimum_component_index{};
 constexpr int32_t maximum_component_index{maximum_component_count - 1};
 constexpr int32_t minimum_bits_per_sample{2};
